+++ test/sensor_simulator/imu.h
@@ -58,7 +58,7 @@ private:
 	Vector3f _accel_data;
 	Vector3f _gyro_data;
 
-	void send(uint32_t time) override;
+	void send(uint64_t time) override;
 
 };
